requirements.txt
ddddocr==1.4.3
numpy==1.22.3
opencv_python==4.5.5.64
Pillow==9.1.0
requests==2.27.1
selenium==4.1.3
